optim: 整理tools
